+++ next.config.mjs
@@ -2,13 +2,20 @@ import createMDX from "@next/mdx";
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  pageExtensions: ["js", "jsx", "md", "mdx", "ts", "tsx"],
+  pageExtensions: ["js", "jsx", "md", "ts", "tsx"],
   transpilePackages: ['next-mdx-remote'],
 };
 
 const withMDX = createMDX({
-  extension: /\.(md|mdx)$/,
-  // plugins go here
+  extension: /\.md$/,
+  options: {
+    remarkPlugins: [
+      "remark-gfm",
+      "remark-math",
+      "remark-frontmatter",
+    ],
+    format: "md",
+  },
 });
 
 export default withMDX(nextConfig);
